Improve UI in NotesList & NoteDetails screens + Add SerializableNote NavType support
