+++ app/src/main/java/academy/apparchitects/notesapp/ui/CustomNavType.kt
@@ -1,0 +1,27 @@
+package academy.apparchitects.notesapp.ui
+
+import android.os.Build
+import android.os.Bundle
+import android.os.Parcelable
+import androidx.navigation.NavType
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+
+inline fun <reified T : Parcelable> parcelableType(
+  isNullableAllowed: Boolean = false,
+  json: Json = Json,
+) = object : NavType<T>(isNullableAllowed = isNullableAllowed) {
+  override fun get(bundle: Bundle, key: String) =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+      bundle.getParcelable(key, T::class.java)
+    } else {
+      @Suppress("DEPRECATION")
+      bundle.getParcelable(key)
+    }
+
+  override fun parseValue(value: String): T = json.decodeFromString(value)
+
+  override fun serializeAsValue(value: T): String = json.encodeToString(value)
+
+  override fun put(bundle: Bundle, key: String, value: T) = bundle.putParcelable(key, value)
+}
