+++ app/src/main/java/academy/apparchitects/notesapp/ui/NotesAppNav.kt
@@ -1,60 +1,86 @@
 package academy.apparchitects.notesapp.ui
 
-import academy.apparchitects.notesapp.ui.screens.NotesListScreen
+import academy.apparchitects.notesapp.data.SerializableNote
 import academy.apparchitects.notesapp.ui.screens.note_details.NoteDetailsScreen
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
+import academy.apparchitects.notesapp.ui.screens.note_list.NotesListScreen
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import kotlinx.serialization.Serializable
-import kotlin.random.Random
-
+import kotlin.reflect.typeOf
 
 @Composable
 fun NotesAppNav(
-    navController: NavHostController,
-    startDestination: Destinations = Destinations.NotesList
+  navController: NavHostController,
+  startDestination: Destinations = Destinations.NotesList
 ) {
 
-    NavHost(
-        navController = navController,
-        startDestination = startDestination
-    ) {
+  NavHost(
+    navController = navController,
+    startDestination = startDestination
+  ) {
 
-        composable<Destinations.NotesList> {
-            NotesListScreen(onNoteClick = {
-                navController.navigate(Destinations.NoteDetail(it))
-            },
-              onAddNoteClick = {
-                navController.navigate(Destinations.NoteDetail())
-              }
-            )
+    composable<Destinations.NotesList> {
+      NotesListScreen(onNoteClick = { noteId, note ->
+//        navController.navigate(
+//          Destinations.NoteDetail(noteId)
+//        )
+        // TODO: This is hacky for now to pass the note object to the details screen
+        // Change it once repository layer is there
+        navController.navigate(
+          Destinations.SerializableNoteDetail(
+            noteId = noteId,
+            note = note.toSerializableNote()
+          )
+        )
+      },
+        onAddNoteClick = {
+          navController.navigate(Destinations.NoteDetail())
         }
-
-        composable<Destinations.NoteDetail> { backStackEntry ->
-            val arg: Destinations.NoteDetail = backStackEntry.toRoute()
-
-            NoteDetailsScreen(
-              navigateUp = {navController.navigateUp()},
-              onShareClick = { _ ->
-
-              },
-              noteId = arg.noteId
-            )
-        }
-
+      )
     }
+
+    composable<Destinations.NoteDetail> { backStackEntry ->
+      val args: Destinations.NoteDetail = backStackEntry.toRoute()
+
+      NoteDetailsScreen(
+        navigateUp = { navController.navigateUp() },
+        onShareClick = { _ -> },
+        noteId = args.noteId
+      )
+    }
+
+    composable<Destinations.SerializableNoteDetail>(
+      typeMap = mapOf(typeOf<SerializableNote>() to parcelableType<SerializableNote>())
+    ) { backStackEntry ->
+      val args: Destinations.SerializableNoteDetail = backStackEntry.toRoute()
+
+      NoteDetailsScreen(
+        navigateUp = { navController.navigateUp() },
+        onShareClick = { _ -> },
+        noteId = args.noteId,
+        serializableNote = args.note
+      )
+    }
+
+  }
 }
 
 sealed class Destinations {
 
-    @Serializable
-    object NotesList: Destinations()
+  @Serializable
+  data object NotesList : Destinations()
 
-    @Serializable
-    data class NoteDetail(val noteId: String? = null): Destinations()
+  @Serializable
+  data class NoteDetail(
+    val noteId: String? = null
+  ) : Destinations()
+
+  @Serializable
+  data class SerializableNoteDetail(
+    val noteId: String? = null,
+    val note: SerializableNote
+  ) : Destinations()
 }
