+++ data/build.gradle.kts
@@ -1,6 +1,8 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.jetbrains.kotlin.android)
+    kotlin("plugin.serialization").version("1.9.0")
+    id("kotlin-parcelize")
 }
 
 android {
@@ -41,4 +43,6 @@ dependencies {
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
     implementation(libs.kotlinx.datetime)
+
+    implementation(libs.kotlinx.serialization.json)
 }
